Bring melodic-devel closer to ros2 branch (#76)
* comply with extra and pedantic compiler flags

* use c++11 nullptr instead of NULL

* make ABI breaking change for explicit constructors

* make linters happy

* no need to support console_bridge < 0.3.0 anymore

* remove obsolete todo

* add virtual destructor in test

* vector size() returns size_t

* simplify branching

+++ test/unique_ptr_test.cpp
@@ -34,6 +34,7 @@
 #include <gtest/gtest.h>
 #include <boost/thread.hpp>
 
+#include <cstddef>
 #include <functional>
 #include <iostream>
 #include <string>
@@ -46,7 +47,6 @@ const char LIBRARY_2[] = "libclass_loader_TestPlugins2.so";
 
 using class_loader::ClassLoader;
 
-/*****************************************************************************/
 TEST(ClassLoaderUniquePtrTest, basicLoad) {
   try {
     ClassLoader loader1(LIBRARY_1, false);
@@ -57,7 +57,6 @@ TEST(ClassLoaderUniquePtrTest, basicLoad) {
   }
 }
 
-/*****************************************************************************/
 TEST(ClassLoaderUniquePtrTest, correctLazyLoadUnload) {
   try {
     ASSERT_FALSE(class_loader::class_loader_private::isLibraryLoadedByAnybody(LIBRARY_1));
@@ -81,8 +80,6 @@ TEST(ClassLoaderUniquePtrTest, correctLazyLoadUnload) {
   }
 }
 
-/*****************************************************************************/
-
 TEST(ClassLoaderUniquePtrTest, nonExistentPlugin) {
   ClassLoader loader1(LIBRARY_1, false);
 
@@ -103,8 +100,6 @@ TEST(ClassLoaderUniquePtrTest, nonExistentPlugin) {
   FAIL() << "Did not throw exception as expected.\n";
 }
 
-/*****************************************************************************/
-
 void wait(int seconds)
 {
   boost::this_thread::sleep(boost::posix_time::seconds(seconds));
@@ -113,7 +108,7 @@ void wait(int seconds)
 void run(ClassLoader * loader)
 {
   std::vector<std::string> classes = loader->getAvailableClasses<Base>();
-  for (unsigned int c = 0; c < classes.size(); c++) {
+  for (size_t c = 0; c < classes.size(); c++) {
     loader->createUniqueInstance<Base>(classes.at(c))->saySomething();
   }
 }
@@ -128,11 +123,11 @@ TEST(ClassLoaderUniquePtrTest, threadSafety) {
   try {
     std::vector<boost::thread> client_threads;
 
-    for (unsigned int c = 0; c < 1000; c++) {
+    for (size_t c = 0; c < 1000; c++) {
       client_threads.emplace_back(std::bind(&run, &loader1));
     }
 
-    for (unsigned int c = 0; c < client_threads.size(); c++) {
+    for (size_t c = 0; c < client_threads.size(); c++) {
       client_threads.at(c).join();
     }
 
@@ -144,9 +139,6 @@ TEST(ClassLoaderUniquePtrTest, threadSafety) {
     FAIL() << "Unknown exception.";
   }
 }
-
-
-/*****************************************************************************/
 
 TEST(ClassLoaderUniquePtrTest, loadRefCountingLazy) {
   try {
@@ -187,9 +179,6 @@ TEST(ClassLoaderUniquePtrTest, loadRefCountingLazy) {
 
   FAIL() << "Did not throw exception as expected.\n";
 }
-
-
-/*****************************************************************************/
 
 void testMultiClassLoader(bool lazy)
 {
@@ -239,8 +228,6 @@ TEST(MultiClassLoaderUniquePtrTest, noWarningOnLazyLoad) {
 
   SUCCEED();
 }
-
-/*****************************************************************************/
 
 // Run all the tests that were declared with TEST()
 int main(int argc, char ** argv)
